external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/canvas/drawingbuffer-static-canvas-test-expected.txt

This test ensures WebGL implementations correctly implement drawingbufferWidth/Height with compositing.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS context exists
PASS gl.program is non-null.
PASS gl.getError() is gl.NO_ERROR

Checking drawingBufferWidth/drawingBufferHeight
PASS gl.drawingBufferWidth is gl.canvas.width
PASS gl.drawingBufferHeight is gl.canvas.height
PASS maxSize[0] > 0 is true
PASS maxSize[1] > 0 is true
PASS gl.drawingBufferWidth > 0 is true
PASS gl.drawingBufferHeight > 0 is true
PASS gl.drawingBufferWidth <= maxSize[0] is true
PASS gl.drawingBufferHeight <= maxSize[1] is true
PASS gl.getError() is gl.NO_ERROR

Checking scaling up then back down to 50/50, drawing still works.
PASS gl.drawingBufferWidth == 50 is true
PASS gl.drawingBufferHeight == 50 is true
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR

PASS successfullyParsed is true

TEST COMPLETE
